requirements.txt
absl-py
accelerate==0.12.0
av 
blobfile 
decord 
einops
ftfy==6.1.1
gdown
git+https://github.com/openai/CLIP.git
huggingface_hub
imageio
ml_collections
moviepy 
natsort
open_clip_torch
pandas
pytorch-fid
regex
sentencepiece 
timm
tqdm
transformers==4.29.2
triton
wandb
xformers